+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Extract service objects in subdirectory
 - Rename `SmartCheckout` facade to `Viva`
-- Implement payment orders for ISV
+- Create payment orders for ISV
+- Retrieve transaction by id for ISV
 
 ## [6.0.0-alpha.1] - 2022-11-08